Disable database migration on unsupported platforms or if the nobolt tag is present
