+++ v2/bolt_store.go
@@ -2,12 +2,9 @@ package raftboltdb
 
 import (
 	"errors"
-	"fmt"
-	"os"
 	"time"
 
 	metrics "github.com/armon/go-metrics"
-	v1 "github.com/boltdb/bolt"
 	"github.com/hashicorp/raft"
 	"go.etcd.io/bbolt"
 )
@@ -25,6 +22,8 @@ var (
 
 	// An error indicating a given key does not exist
 	ErrKeyNotFound = errors.New("not found")
+	// An error indicating a feature isn't implemented on the current platform
+	ErrNotImplemented = errors.New("not implemented")
 )
 
 // BoltStore provides access to Bbolt for Raft to store and retrieve
@@ -296,68 +295,4 @@ func (b *BoltStore) GetUint64(key []byte) (uint64, error) {
 // database file to sync against the disk.
 func (b *BoltStore) Sync() error {
 	return b.conn.Sync()
-}
-
-// MigrateToV2 reads in the source file path of a BoltDB file
-// and outputs all the data migrated to a Bbolt destination file
-func MigrateToV2(source, destination string) (*BoltStore, error) {
-	_, err := os.Stat(destination)
-	if err == nil {
-		return nil, fmt.Errorf("file exists in destination %v", destination)
-	}
-
-	srcDb, err := v1.Open(source, dbFileMode, &v1.Options{
-		ReadOnly: true,
-		Timeout:  1 * time.Minute,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed opening source database: %v", err)
-	}
-
-	//Start a connection to the source
-	srctx, err := srcDb.Begin(false)
-	if err != nil {
-		return nil, fmt.Errorf("failed connecting to source database: %v", err)
-	}
-	defer srctx.Rollback()
-
-	//Create the destination
-	destDb, err := New(Options{Path: destination})
-	if err != nil {
-		return nil, fmt.Errorf("failed creating destination database: %v", err)
-	}
-	//Start a connection to the new
-	desttx, err := destDb.conn.Begin(true)
-	if err != nil {
-		destDb.Close()
-		os.Remove(destination)
-		return nil, fmt.Errorf("failed connecting to destination database: %v", err)
-	}
-
-	defer desttx.Rollback()
-
-	//Loop over both old buckets and set them in the new
-	buckets := [][]byte{dbConf, dbLogs}
-	for _, b := range buckets {
-		srcB := srctx.Bucket(b)
-		destB := desttx.Bucket(b)
-		err = srcB.ForEach(func(k, v []byte) error {
-			return destB.Put(k, v)
-		})
-		if err != nil {
-			destDb.Close()
-			os.Remove(destination)
-			return nil, fmt.Errorf("failed to copy %v bucket: %v", string(b), err)
-		}
-	}
-
-	//If the commit fails, clean up
-	if err := desttx.Commit(); err != nil {
-		destDb.Close()
-		os.Remove(destination)
-		return nil, fmt.Errorf("failed commiting data to destination: %v", err)
-	}
-
-	return destDb, nil
-
 }
